feat: Add data to booksReducer
