+++ src/redux/books/books.js
@@ -13,7 +13,32 @@ export const REMOVE_BOOK = (book) => ({
 
 // Language: javascript
 // Path: src/redux/books/books.js
-const books = [];
+const books = [
+  {
+    id: 1,
+    title: 'The Hunger Games',
+    genres: ['Action', 'Adventure', 'Sci-Fi'],
+    author: 'Suzanne Collins',
+    progress: 64,
+    chapter: 'Chapter 17',
+  },
+  {
+    id: 2,
+    title: 'Dune',
+    genres: ['Sci-Fi', 'Fantasy'],
+    author: 'Frank Herbert',
+    progress: 8,
+    chapter: 'Chapter 3: "A Lesson Learned"',
+  },
+  {
+    id: 3,
+    title: 'Capital in the Twenty-First Century',
+    genres: ['Economics', 'Non-Fiction'],
+    author: 'Suzanne Collins',
+    progress: 0,
+    chapter: 'Introduction',
+  },
+];
 
 const booksReducer = (state = books, action) => {
   switch (action.type) {
